// src/utils/types.ts
export type Product = {
  cart_id: string;
  product_id: string;
  product_name: string;
  product_price: number;
  product_size: string;
  product_quantity: number;
  product_variant_id: string;
  product_variant_size: string;
  product_variant_color: string;
  product_variant_quantity: number;
  product_variant_image: string;
};